fix(providers): 统一两端 Anthropic 上下文窗口裁决到共享真源
WebUI settings 的 findProviderModelConfig/getKnownModelLimits 还是旧的
内联三元逻辑，缺少 baseUrl === undefined 放行分支：[1m] 模型未配 baseUrl
时会用目录默认窗口覆盖用户配置值，导致用量环两端读数不一致（GUI 正常、
WebUI 显示模型默认窗口）。

窗口策略（resolveAnthropicContextWindow/resolveAnthropicKnownModelLimits/
shouldSendAnthropicLongContextHeader 及常量）抽到共享包
agent-ui/lib/models/anthropicContext.ts 作为唯一真源；桌面端
anthropicModels.ts 改为 re-export 并保留请求路径专属逻辑，WebUI settings
直接消费共享实现。补充 [1m] 自定义窗口保留与桌面改值同步生效的回归测试。
